backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "openai-whisper>=20250625",
    "uvicorn[standard]>=0.40.0",
    "youtube-transcript-api>=1.2.3",
    "yt-dlp>=2025.12.8",
]
